ref(replays): cut two type-only edges into the import cycle
Two small, source-side type-only cuts (no behavior change):

- `utils/replays/types` imported `Event` solely to type one type-guard
  param (`isHydrateCrumb`). The guard only reads `.category`, so it now
  takes `unknown` and narrows structurally — dropping the `types/event`
  edge. (~-24 to the largest SCC on its own.)
- The `Output` enum lived in `getOutputType` (a runtime util), so the
  analytics event map `replayAnalyticsEvents` pulled that whole module in
  for one type. Move `Output` to a dependency-free leaf and re-export it
  from `getOutputType` for existing consumers. (~-6 on its own.)

Largest type-import SCC drops 1681 -> 1651 (-30) across 5 files.

Note: the sibling 'cheap cut' candidates (uptime/types, deprecatedforms,
issueTypeConfig importing core/group types) are relocation-style and are
already handled by the leaf-extraction PR; configStore genuinely stores
`Config` and isn't cheaply cuttable.

Part of a series of SCC-shrinking PRs.
