+++ static/app/utils/replays/types.tsx
@@ -13,8 +13,6 @@ import type {
 } from '@sentry/react';
 import invariant from 'invariant';
 
-import type {Event} from 'sentry/types/event';
-
 export type {serializedNodeWithId} from '@sentry-internal/rrweb-snapshot';
 export type {fullSnapshotEvent, incrementalSnapshotEvent} from '@sentry-internal/rrweb';
 
@@ -214,8 +212,13 @@ export function isFeedbackFrame(frame: ReplayFrame | undefined): frame is Feedba
   return Boolean(frame && 'category' in frame && frame.category === 'feedback');
 }
 
-export function isHydrateCrumb(item: BreadcrumbFrame | Event): item is BreadcrumbFrame {
-  return 'category' in item && item.category === 'replay.hydrate-error';
+export function isHydrateCrumb(item: unknown): item is BreadcrumbFrame {
+  return (
+    typeof item === 'object' &&
+    item !== null &&
+    'category' in item &&
+    item.category === 'replay.hydrate-error'
+  );
 }
 
 export function isSpanFrame(frame: ReplayFrame | undefined): frame is SpanFrame {
