+++ static/app/views/explore/replays/detail/network/details/output.tsx
@@ -1,0 +1,10 @@
+export enum Output {
+  SETUP = 'setup',
+  UNSUPPORTED = 'unsupported',
+  URL_SKIPPED = 'url_skipped',
+  BODY_SKIPPED = 'body_skipped',
+  BODY_PARSE_ERROR = 'body_parse_error',
+  BODY_PARSE_TIMEOUT = 'body_parse_timeout',
+  UNPARSEABLE_BODY_TYPE = 'unparseable_body_type',
+  DATA = 'data',
+}
